requirements.txt
datasets>=2.14.0
pandas>=1.5.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
python-dotenv>=0.19.0
pydantic>=2.0.0
PyYAML>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
wordcloud>=1.9.0
plotly>=5.0.0
kaleido>=0.2.0

# Deep Learning & NLP for Emotiment Module
transformers>=4.30.0
torch>=2.0.0
peft>=0.4.0
accelerate>=0.20.0
evaluate>=0.4.0
scikit-learn>=1.3.0
xgboost>=1.7.0
emoji>=2.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
